requirements.txt
setuptools-rust~=0.11.6
ccxt~=1.41.45
pyyaml~=5.4.1
PyJWT==1.7.1
python-jose~=3.2.0
sanic~=21.3.4
uvloop~=0.15.2
httptools~=0.2.0
jinja2~=2.11.3